perf:优化count查询

+++ src/main/java/io/github/heathchen/mybatisplus/util/domain/CountBuilder.java
@@ -96,7 +96,7 @@ public class CountBuilder<E> {
         return this;
     }
 
-    public Integer doCountByReflect() {
+    public Long doCountByReflect() {
         return MyBatisPlusUtil.countByReflect(this);
     }
 }
